+++ Analysis/Tasks/trackselection.cxx
@@ -44,8 +44,8 @@ struct TrackSelectionTask {
   void process(soa::Join<aod::FullTracks, aod::TracksExtended> const& tracks)
   {
     for (auto& track : tracks) {
-      filterTable(globalTracks.IsSelected(track),
-                  globalTracksSDD.IsSelected(track));
+      filterTable((uint8_t)globalTracks.IsSelected(track),
+                  (uint8_t)globalTracksSDD.IsSelected(track));
     }
   }
 };
